data_services/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
bs4
selenium
pandas
tqdm
sqlalchemy
pymysql
